backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.4
plaid-python==9.1.0
python-dotenv==1.0.0
